Solution: Arrays 101. Find Numbers with Even Number of Digits
This commit adds a spec and solution file for the Find Numbers with Even
Number of Digits challenge. This toy problem asks that you write a function that
accepts a single Array parameter.  The function must return the number of
numbers in the Array that have an even number of digits.