Solution: Arrays 101. Squares of a Sorted Array
This commit adds a spec and solution file for the Squares of a Sorted
Array challenge. This toy problem asks that you write a function that
accepts a single Array parameter of numbers sorted in ascending order and
return an Array of the squares of the numbers in the input array, also sorted in
ascending order.