Solution: Arrays 101. Duplicate Zeros
This commit adds a spec and solution file for the Duplicate Zeros
challenge. This was the first challenge that I submitted that was
incorrect. I missed an important part of the instructions and thought
the length of the array was supposed to increase (as necessary) to
include the added zeros. This was not the case. The length of the array
was not supposed to change.